fix bugs while integration Testing of migrate data
