+++ cmd/cli/command/shardadd.go
@@ -7,9 +7,10 @@ import (
 	"sort"
 	"time"
 	"strings"
+	"strconv"
 
 	"gopkg.in/urfave/cli.v1"
-	cliCtx "github.com/KvrocksLabs/kvrocks-controller/cmd/cli/context"
+	"github.com/KvrocksLabs/kvrocks-controller/cmd/cli/context"
 	"github.com/KvrocksLabs/kvrocks-controller/util"
 	"github.com/KvrocksLabs/kvrocks-controller/server/handlers"
 )
@@ -42,8 +43,8 @@ var AddShardCommand = cli.Command{
 }
 
 func addShardAction(c *cli.Context) {
-	ctx := cliCtx.GetContext()
-	if ctx.Location != cliCtx.LocationCluster {
+	ctx := context.GetContext()
+	if ctx.Location != context.LocationCluster {
 		fmt.Println("mkcl command should under clsuter dir")
 		return 
 	}
@@ -115,65 +116,13 @@ func addShardAction(c *cli.Context) {
 				shardParam.Slaves = shard.Nodes[1:]
 			}
 			resp, err := util.HttpPost(handlers.GetShardRootURL(ctx.Leader, ctx.Namespace, ctx.Cluster), shardParam, 5 * time.Second)
-			if err != nil {
-				fmt.Println("create shard ", idx, " error: " + err.Error())
-				return 
-			}
-			if resp.Errno != handlers.Success {
-				fmt.Println("create shard  ", idx, " error: " + resp.Errmsg)	
+			if HttpResponeException("creare shard" + strconv.Itoa(idx), resp, err) {
 				return
 			}
-			if resp.Body == nil {
-				fmt.Println("create shard  ", idx, "error")
-				return
-			}
-			fmt.Println("crate shard: ", idx, "response: ", resp.Body.(string))
+			fmt.Println("crate shard", idx, "response: ", resp.Body.(string))
 		}
 	} else {
 		fmt.Println("add -d param, do above make shard plan")
 	}
 	return 
-}
-
-var DelShardCommand = cli.Command{
-	Name:      "delshard",
-	Usage:     "del shard",
-	ArgsUsage: "-i ${shard_idx}",
-	Action:    delShardAction,
-	Flags: []cli.Flag{
-		cli.IntFlag{
-			Name:  "i,shardidx", 
-			Value: -1, 
-			Usage: "shard number"},
-	},
-	Description: `
-    del shard under special cluster
-    `,
-}
-
-func delShardAction(c *cli.Context) {
-	ctx := cliCtx.GetContext()
-	if ctx.Location != cliCtx.LocationCluster {
-		fmt.Println("mkcl command should under clsuter dir")
-		return 
-	}
-	shardIdx := c.Int("i")
-	if shardIdx < 0 {
-		fmt.Println("shard_idx(-i) error")
-		return
-	}
-	resp, err := util.HttpDelete(handlers.GetShardURL(ctx.Leader, ctx.Namespace, ctx.Cluster, shardIdx), nil, 5 * time.Second)
-	if err != nil {
-		fmt.Println("delete shard error: " + err.Error())
-		return 
-	}
-	if resp.Errno != handlers.Success {
-		fmt.Println("create shard error: " + resp.Errmsg)	
-		return
-	}
-	if resp.Body == nil {
-		fmt.Println("create shard error")
-		return
-	}
-	fmt.Println(resp.Body.(string))
 }
