+++ consts/context_key.go
@@ -2,4 +2,5 @@ package consts
 
 const (
 	ContextKeyStorage = "_context_key_storage"
+	ContextKeyMigrate = "_context_key_migrate"
 )
