+++ server/handlers/cluster.go
@@ -8,6 +8,7 @@ import (
 	"github.com/KvrocksLabs/kvrocks-controller/consts"
 	"github.com/KvrocksLabs/kvrocks-controller/metadata"
 	"github.com/KvrocksLabs/kvrocks-controller/storage"
+	"github.com/KvrocksLabs/kvrocks-controller/util"
 	"github.com/gin-gonic/gin"
 )
 
@@ -29,13 +30,13 @@ func ListCluster(c *gin.Context) {
 	clusters, err := stor.ListCluster(namespace)
 	if err != nil {
 		if metaErr, ok := err.(*metadata.Error); ok && metaErr.Code == metadata.CodeNoExists {
-			c.JSON(http.StatusNotFound, MakeFailureResponse(err.Error()))
+			c.JSON(http.StatusNotFound, util.MakeFailureResponse(err.Error()))
 		} else {
-			c.JSON(http.StatusInternalServerError, MakeFailureResponse(err.Error()))
+			c.JSON(http.StatusInternalServerError, util.MakeFailureResponse(err.Error()))
 		}
 		return
 	}
-	c.JSON(http.StatusOK, MakeSuccessResponse(clusters))
+	c.JSON(http.StatusOK, util.MakeSuccessResponse(clusters))
 }
 
 func GetCluster(c *gin.Context) {
@@ -45,13 +46,13 @@ func GetCluster(c *gin.Context) {
 	cluster, err := stor.GetClusterCopy(namespace, clusterName)
 	if err != nil {
 		if metaErr, ok := err.(*metadata.Error); ok && metaErr.Code == metadata.CodeNoExists {
-			c.JSON(http.StatusNotFound, MakeFailureResponse(err.Error()))
+			c.JSON(http.StatusNotFound, util.MakeFailureResponse(err.Error()))
 		} else {
-			c.JSON(http.StatusInternalServerError, MakeFailureResponse(err.Error()))
+			c.JSON(http.StatusInternalServerError, util.MakeFailureResponse(err.Error()))
 		}
 		return
 	}
-	c.JSON(http.StatusOK, MakeSuccessResponse(cluster))
+	c.JSON(http.StatusOK, util.MakeSuccessResponse(cluster))
 }
 
 func CreateCluster(c *gin.Context) {
@@ -60,11 +61,11 @@ func CreateCluster(c *gin.Context) {
 
 	var req CreateClusterParam
 	if err := c.BindJSON(&req); err != nil {
-		c.JSON(http.StatusBadRequest, MakeFailureResponse(err.Error()))
+		c.JSON(http.StatusBadRequest, util.MakeFailureResponse(err.Error()))
 		return
 	}
 	if err := req.validate(); err != nil {
-		c.JSON(http.StatusBadRequest, MakeFailureResponse(err.Error()))
+		c.JSON(http.StatusBadRequest, util.MakeFailureResponse(err.Error()))
 		return
 	}
 	shards := make([]metadata.Shard, len(req.Shards))
@@ -72,7 +73,7 @@ func CreateCluster(c *gin.Context) {
 	for i, createShard := range req.Shards {
 		shard, err := createShard.toShard()
 		if err != nil {
-			c.JSON(http.StatusBadRequest, MakeFailureResponse("index: " + strconv.Itoa(i) + ", err: " + err.Error()))
+			c.JSON(http.StatusBadRequest, util.MakeFailureResponse("index: " + strconv.Itoa(i) + ", err: " + err.Error()))
 			return
 		}
 		shard.SlotRanges = append(shard.SlotRanges, slotRanges[i])
@@ -81,13 +82,13 @@ func CreateCluster(c *gin.Context) {
 
 	if err := stor.CreateCluster(namespace, req.Cluster, &metadata.Cluster{Shards: shards, }); err != nil {
 		if metaErr, ok := err.(*metadata.Error); ok && metaErr.Code == metadata.CodeExisted {
-			c.JSON(http.StatusConflict, MakeFailureResponse(err.Error()))
+			c.JSON(http.StatusConflict, util.MakeFailureResponse(err.Error()))
 		} else {
-			c.JSON(http.StatusInternalServerError, MakeFailureResponse(err.Error()))
+			c.JSON(http.StatusInternalServerError, util.MakeFailureResponse(err.Error()))
 		}
 		return
 	}
-	c.JSON(http.StatusCreated, MakeSuccessResponse("OK"))
+	c.JSON(http.StatusCreated, util.MakeSuccessResponse("OK"))
 }
 
 func RemoveCluster(c *gin.Context) {
@@ -96,11 +97,11 @@ func RemoveCluster(c *gin.Context) {
 	cluster := c.Param("cluster")
 	if err := stor.RemoveCluster(namespace, cluster); err != nil {
 		if metaErr, ok := err.(*metadata.Error); ok && metaErr.Code == metadata.CodeNoExists {
-			c.JSON(http.StatusNotFound, MakeFailureResponse(err.Error()))
+			c.JSON(http.StatusNotFound, util.MakeFailureResponse(err.Error()))
 		} else {
-			c.JSON(http.StatusInternalServerError, MakeFailureResponse(err.Error()))
+			c.JSON(http.StatusInternalServerError, util.MakeFailureResponse(err.Error()))
 		}
 		return
 	}
-	c.JSON(http.StatusOK, MakeSuccessResponse("OK"))
+	c.JSON(http.StatusOK, util.MakeSuccessResponse("OK"))
 }
