+++ server/handlers/controller.go
@@ -5,16 +5,17 @@ import (
 
 	"github.com/KvrocksLabs/kvrocks-controller/consts"
 	"github.com/KvrocksLabs/kvrocks-controller/storage"
+	"github.com/KvrocksLabs/kvrocks-controller/util"
 	"github.com/gin-gonic/gin"
 )
 
 func Leader(c *gin.Context) {
 	stor := c.MustGet(consts.ContextKeyStorage).(*storage.Storage)
-	c.JSON(http.StatusOK, MakeSuccessResponse(stor.Leader()))
+	c.JSON(http.StatusOK, util.MakeSuccessResponse(stor.Leader()))
 }
 
-func ReleaseLeader(c *gin.Context) {
+func LeaderResign(c *gin.Context) {
 	stor := c.MustGet(consts.ContextKeyStorage).(*storage.Storage)
 	stor.LeaderResign()
-	c.JSON(http.StatusOK, MakeSuccessResponse("OK"))
+	c.JSON(http.StatusOK, util.MakeSuccessResponse("OK"))
 }
