+++ server/handlers/shard.go
@@ -8,6 +8,8 @@ import (
 	"github.com/KvrocksLabs/kvrocks-controller/consts"
 	"github.com/KvrocksLabs/kvrocks-controller/metadata"
 	"github.com/KvrocksLabs/kvrocks-controller/storage"
+	"github.com/KvrocksLabs/kvrocks-controller/migrate"
+	"github.com/KvrocksLabs/kvrocks-controller/util"
 	"github.com/gin-gonic/gin"
 )
 
@@ -53,13 +55,13 @@ func ListShard(c *gin.Context) {
 	shards, err := stor.ListShard(ns, cluster)
 	if err != nil {
 		if metaErr, ok := err.(*metadata.Error); ok && metaErr.Code == metadata.CodeNoExists {
-			c.JSON(http.StatusNotFound, MakeFailureResponse(err.Error()))
+			c.JSON(http.StatusNotFound, util.MakeFailureResponse(err.Error()))
 		} else {
-			c.JSON(http.StatusInternalServerError, MakeFailureResponse(err.Error()))
+			c.JSON(http.StatusInternalServerError, util.MakeFailureResponse(err.Error()))
 		}
 		return
 	}
-	c.JSON(http.StatusOK, MakeSuccessResponse(shards))
+	c.JSON(http.StatusOK, util.MakeSuccessResponse(shards))
 }
 
 func GetShard(c *gin.Context) {
@@ -67,7 +69,7 @@ func GetShard(c *gin.Context) {
 	cluster := c.Param("cluster")
 	shard, err := strconv.Atoi(c.Param("shard"))
 	if err != nil {
-		c.JSON(http.StatusBadRequest, MakeFailureResponse(err.Error()))
+		c.JSON(http.StatusBadRequest, util.MakeFailureResponse(err.Error()))
 		return
 	}
 
@@ -75,13 +77,13 @@ func GetShard(c *gin.Context) {
 	s, err := stor.GetShard(ns, cluster, shard)
 	if err != nil {
 		if metaErr, ok := err.(*metadata.Error); ok && metaErr.Code == metadata.CodeNoExists {
-			c.JSON(http.StatusNotFound, MakeFailureResponse(err.Error()))
+			c.JSON(http.StatusNotFound, util.MakeFailureResponse(err.Error()))
 		} else {
-			c.JSON(http.StatusInternalServerError, MakeFailureResponse(err.Error()))
+			c.JSON(http.StatusInternalServerError, util.MakeFailureResponse(err.Error()))
 		}
 		return
 	}
-	c.JSON(http.StatusOK, MakeSuccessResponse(s))
+	c.JSON(http.StatusOK, util.MakeSuccessResponse(s))
 }
 
 func CreateShard(c *gin.Context) {
@@ -90,25 +92,25 @@ func CreateShard(c *gin.Context) {
 
 	var req CreateShardParam
 	if err := c.BindJSON(&req); err != nil {
-		c.JSON(http.StatusBadRequest, MakeFailureResponse(err.Error()))
+		c.JSON(http.StatusBadRequest, util.MakeFailureResponse(err.Error()))
 		return
 	}
 	shard, err := req.toShard()
 	if err != nil {
-		c.JSON(http.StatusBadRequest, MakeFailureResponse(err.Error()))
+		c.JSON(http.StatusBadRequest, util.MakeFailureResponse(err.Error()))
 		return
 	}
 
 	stor := c.MustGet(consts.ContextKeyStorage).(*storage.Storage)
 	if err := stor.CreateShard(ns, cluster, shard); err != nil {
 		if metaErr, ok := err.(*metadata.Error); ok && metaErr.Code == metadata.CodeExisted {
-			c.JSON(http.StatusConflict, MakeFailureResponse(err.Error()))
+			c.JSON(http.StatusConflict, util.MakeFailureResponse(err.Error()))
 		} else {
-			c.JSON(http.StatusInternalServerError, MakeFailureResponse(err.Error()))
+			c.JSON(http.StatusInternalServerError, util.MakeFailureResponse(err.Error()))
 		}
 		return
 	}
-	c.JSON(http.StatusCreated, MakeSuccessResponse("OK"))
+	c.JSON(http.StatusCreated, util.MakeSuccessResponse("OK"))
 }
 
 func RemoveShard(c *gin.Context) {
@@ -116,20 +118,20 @@ func RemoveShard(c *gin.Context) {
 	cluster := c.Param("cluster")
 	shard, err := strconv.Atoi(c.Param("shard"))
 	if err != nil {
-		c.JSON(http.StatusBadRequest, MakeFailureResponse(err.Error()))
+		c.JSON(http.StatusBadRequest, util.MakeFailureResponse(err.Error()))
 		return
 	}
 
 	stor := c.MustGet(consts.ContextKeyStorage).(*storage.Storage)
 	if err := stor.RemoveShard(ns, cluster, shard); err != nil {
 		if metaErr, ok := err.(*metadata.Error); ok && metaErr.Code == metadata.CodeNoExists {
-			c.JSON(http.StatusNotFound, MakeFailureResponse(err.Error()))
+			c.JSON(http.StatusNotFound, util.MakeFailureResponse(err.Error()))
 		} else {
-			c.JSON(http.StatusInternalServerError, MakeFailureResponse(err.Error()))
+			c.JSON(http.StatusInternalServerError, util.MakeFailureResponse(err.Error()))
 		}
 		return
 	}
-	c.JSON(http.StatusCreated, MakeSuccessResponse("OK"))
+	c.JSON(http.StatusCreated, util.MakeSuccessResponse("OK"))
 }
 
 func UpdateShardSlots(c *gin.Context) {
@@ -138,19 +140,19 @@ func UpdateShardSlots(c *gin.Context) {
 	cluster := c.Param("cluster")
 	shard, err := strconv.Atoi(c.Param("shard"))
 	if err != nil {
-		c.JSON(http.StatusBadRequest, MakeFailureResponse(err.Error()))
+		c.JSON(http.StatusBadRequest, util.MakeFailureResponse(err.Error()))
 		return
 	}
 	var payload ShardSlotsParam
 	if err := c.BindJSON(&payload); err != nil {
-		c.JSON(http.StatusBadRequest, MakeFailureResponse(err.Error()))
+		c.JSON(http.StatusBadRequest, util.MakeFailureResponse(err.Error()))
 		return
 	}
 	slotRanges := make([]metadata.SlotRange, len(payload.Slots))
 	for i, slot := range payload.Slots {
 		slotRange, err := metadata.ParseSlotRange(slot)
 		if err != nil {
-			c.JSON(http.StatusBadRequest, MakeFailureResponse(err.Error()))
+			c.JSON(http.StatusBadRequest, util.MakeFailureResponse(err.Error()))
 			return
 		}
 		slotRanges[i] = *slotRange
@@ -164,11 +166,46 @@ func UpdateShardSlots(c *gin.Context) {
 	}
 	if err != nil {
 		if metaErr, ok := err.(*metadata.Error); ok && metaErr.Code == metadata.CodeNoExists {
-			c.JSON(http.StatusNotFound, MakeFailureResponse(err.Error()))
+			c.JSON(http.StatusNotFound, util.MakeFailureResponse(err.Error()))
 		} else {
-			c.JSON(http.StatusInternalServerError, MakeFailureResponse(err.Error()))
+			c.JSON(http.StatusInternalServerError, util.MakeFailureResponse(err.Error()))
 		}
 		return
 	}
-	c.JSON(http.StatusOK, MakeSuccessResponse("OK"))
+	c.JSON(http.StatusOK, util.MakeSuccessResponse("OK"))
+}
+
+func MigrateSlotsAndData(c *gin.Context) {
+	var migTasks MigrateSlotsDataParam
+	if err := c.BindJSON(&migTasks); err != nil {
+		c.JSON(http.StatusBadRequest, util.MakeFailureResponse(err.Error()))
+		return
+	}
+	migr := c.MustGet(consts.ContextKeyMigrate).(*migrate.Migrate)
+	err := migr.AddMigrateTasks(migTasks.Tasks)
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, util.MakeFailureResponse(err.Error()))
+		return
+	}
+	c.JSON(http.StatusOK, util.MakeSuccessResponse("OK"))
+}
+
+func MigrateSlots(c *gin.Context) {
+	var param MigrateSlotsParam
+	if err := c.BindJSON(&param); err != nil {
+		c.JSON(http.StatusBadRequest, util.MakeFailureResponse(err.Error()))
+		return
+	}
+	ns := c.Param("namespace")
+	cluster := c.Param("cluster")
+	stor := c.MustGet(consts.ContextKeyStorage).(*storage.Storage)
+	if err := stor.RemoveShardSlots(ns, cluster, param.SourceShardIdx, param.SlotRanges); err != nil {
+		c.JSON(http.StatusInternalServerError, util.MakeFailureResponse(err.Error()))
+		return 
+	}
+	if err := stor.AddShardSlots(ns, cluster, param.TargetShardIdx, param.SlotRanges); err != nil {
+		c.JSON(http.StatusInternalServerError, util.MakeFailureResponse(err.Error()))
+		return
+	}
+	c.JSON(http.StatusOK, util.MakeSuccessResponse("OK"))
 }
